+++ home_altitude.cpp
@@ -15,8 +15,8 @@ void home_altitude::mission_start() {
   in_progress_ = true;
 }
 
-void home_altitude::set_altitude(double drone_altitude, double mission_altitude,
-                                 double home_altitude) {
+void home_altitude::set_altitude(float drone_altitude, float mission_altitude,
+                                 float home_altitude) {
   BOOST_VERIFY(in_progress_);
 
   if (home_altitude_.has_value()) {
@@ -25,11 +25,11 @@ void home_altitude::set_altitude(double drone_altitude, double mission_altitude,
     home_altitude_ = home_altitude;
   }
 
-  const double diff{home_altitude_.value() + mission_altitude - drone_altitude};
-  BOOST_VERIFY(std::abs(diff) < 1.0);
+  const float diff{home_altitude_.value() + mission_altitude - drone_altitude};
+  BOOST_VERIFY(std::abs(diff) < 1.0F);
 }
 
-auto home_altitude::get_home_altitude() const -> double {
+auto home_altitude::get_home_altitude() const -> float {
   BOOST_VERIFY(in_progress_);
   BOOST_VERIFY(home_altitude_.has_value());
